backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-multipart==0.0.20
pillow==11.3.0
requests==2.32.4
anthropic==0.67.0
python-dotenv==1.1.1
